shajders

+++ graphics/src/main/resources/pl/warp/engine/graphics/program/util/flags.glsl
@@ -1,5 +1,5 @@
 bool isSet(uint flags, int index) {
-    return ((flags >> index) & 0x01) == 1;
+    return ((flags >> index) & uint(0x01)) == uint(1);
 }
 
 int flag(bool flag, int index){
